previously forgot to actually request perms for notifications, whoops! + call to UserDefaults.standard.synchronize() is unnecessary
