+++ MyNSB/ViewController.swift
@@ -97,14 +97,4 @@ class ViewController: UIViewController {
         }
     }
     
-    @objc func requestNotifications() {
-        let center = UNUserNotificationCenter.current()
-        center.requestAuthorization(options: [.badge, .alert, .sound]) { (Success, Error ) in
-            if Success {
-                print("notifications enabled")
-            } else {
-                print("notification permissions request failed with error: \(Error!)")
-            }
-        }
-    }
 }
